Refactor client form model validation messages and add missing translations (#1153)
