+++ apps/server/AliasVault.Client/Auth/Models/LoginFormModel.cs
@@ -9,36 +9,26 @@ namespace AliasVault.Client.Auth.Models;
 
 using System.ComponentModel.DataAnnotations;
 using AliasVault.Client.Resources;
-using AliasVault.Shared.Models.WebApi.Auth;
 
 /// <summary>
 /// Login form model with localized validation.
 /// </summary>
-public class LoginFormModel : LoginModel
+public class LoginFormModel
 {
     /// <summary>
     /// Gets or sets the username.
     /// </summary>
     [Required(ErrorMessageResourceType = typeof(ValidationMessages), ErrorMessageResourceName = nameof(ValidationMessages.UsernameRequired))]
-    public new string Username { get; set; } = null!;
+    public string Username { get; set; } = null!;
 
     /// <summary>
     /// Gets or sets the password.
     /// </summary>
     [Required(ErrorMessageResourceType = typeof(ValidationMessages), ErrorMessageResourceName = nameof(ValidationMessages.PasswordRequired))]
-    public new string Password { get; set; } = null!;
+    public string Password { get; set; } = null!;
 
     /// <summary>
-    /// Converts the form model to the base model.
+    /// Gets or sets a value indicating whether the user wants to be remembered.
     /// </summary>
-    /// <returns>The base LoginModel.</returns>
-    public LoginModel ToBaseModel()
-    {
-        return new LoginModel
-        {
-            Username = Username,
-            Password = Password,
-            RememberMe = RememberMe,
-        };
-    }
+    public bool RememberMe { get; set; } = true;
 }
