+++ apps/server/AliasVault.Client/Auth/Pages/Unlock.razor
@@ -5,6 +5,7 @@
 @layout Auth.Layout.MainLayout
 @using System.Text.Json
 @using AliasVault.Client.Auth.Components
+@using AliasVault.Client.Auth.Models
 @using AliasVault.Client.Utilities
 @using AliasVault.Shared.Models.WebApi.Auth
 @using AliasVault.Cryptography.Client
