+++ apps/server/AliasVault.Client/Resources/Components/Auth/Login.en.resx
@@ -58,7 +58,7 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
-  
+
   <!-- Page titles -->
   <data name="PageTitle" xml:space="preserve">
     <value>Log in to AliasVault</value>
@@ -72,7 +72,7 @@
     <value>Recovery code verification</value>
     <comment>Title for recovery code step</comment>
   </data>
-  
+
   <!-- Form labels -->
   <data name="UsernameOrEmailLabel" xml:space="preserve">
     <value>Your username or email</value>
@@ -98,7 +98,7 @@
     <value>Remember this machine</value>
     <comment>Label for remember machine checkbox</comment>
   </data>
-  
+
   <!-- Buttons -->
   <data name="LoginButton" xml:space="preserve">
     <value>Log in</value>
@@ -108,7 +108,7 @@
     <value>Login to your account</value>
     <comment>Extended login button text</comment>
   </data>
-  
+
   <!-- Links -->
   <data name="LostPasswordLink" xml:space="preserve">
     <value>Lost Password?</value>
@@ -119,14 +119,14 @@
     <comment>Link text for creating a new vault</comment>
   </data>
   <data name="LoginWithRecoveryCodeLink" xml:space="preserve">
-    <value>log in with a recovery code</value>
+    <value>Log in with a recovery code instead.</value>
     <comment>Link text for logging in with recovery code</comment>
   </data>
   <data name="LoginWithAuthenticatorLink" xml:space="preserve">
-    <value>log in with authenticator code</value>
+    <value>Log in with an authenticator code instead.</value>
     <comment>Link text for logging in with authenticator</comment>
   </data>
-  
+
   <!-- Descriptions and help text -->
   <data name="TwoFactorAuthenticationDescription" xml:space="preserve">
     <value>Your login is protected with an authenticator app. Enter your authenticator code below.</value>
@@ -141,14 +141,14 @@
     <comment>Text before registration link</comment>
   </data>
   <data name="DontHaveAuthenticatorText" xml:space="preserve">
-    <value>Don't have access to your authenticator device? You can</value>
-    <comment>Text before recovery code link</comment>
+    <value>Don't have access to your authenticator device?</value>
+    <comment>Question text about not having access to authenticator</comment>
   </data>
   <data name="RegainedAccessText" xml:space="preserve">
-    <value>Regained access to your authenticator device? You can</value>
-    <comment>Text before authenticator login link</comment>
+    <value>Regained access to your authenticator device?</value>
+    <comment>Question text about regaining access to authenticator</comment>
   </data>
-  
+
   <!-- Loading and status messages -->
   <data name="LoggingInMessage" xml:space="preserve">
     <value>Logging in...</value>
@@ -162,7 +162,7 @@
     <value>Verifying recovery code...</value>
     <comment>Loading message during recovery code verification</comment>
   </data>
-  
+
   <!-- Error messages -->
   <data name="LoginErrorMessage" xml:space="preserve">
     <value>An error occurred while processing the login request. Try again (later).</value>
@@ -172,7 +172,7 @@
     <value>An error occurred while processing the login request.</value>
     <comment>Login request processing error message</comment>
   </data>
-  
+
   <!-- Placeholders -->
   <data name="UsernamePlaceholder" xml:space="preserve">
     <value>name / name@company.com</value>
