+++ apps/server/AliasVault.Client/wwwroot/css/tailwind.css
@@ -2431,11 +2431,6 @@ video {
   background-color: rgb(251 203 116 / var(--tw-bg-opacity));
 }
 
-.hover\:bg-primary-400:hover {
-  --tw-bg-opacity: 1;
-  background-color: rgb(246 167 82 / var(--tw-bg-opacity));
-}
-
 .hover\:bg-primary-600:hover {
   --tw-bg-opacity: 1;
   background-color: rgb(214 131 56 / var(--tw-bg-opacity));
